refactor common merge functions
